+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,48 +1,15 @@
 package roomescape.controller;
 
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.*;
-import roomescape.dto.ReservationResponse;
-import roomescape.dto.ReservationSaveRequest;
-import roomescape.service.ReservationService;
-
-import java.net.URI;
-import java.util.List;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
+@RequestMapping("/reservation")
 public class ReservationController {
 
-    private final ReservationService reservationService;
-
-    public ReservationController(ReservationService reservationService) {
-        this.reservationService = reservationService;
-    }
-
-    @GetMapping("/reservation")
+    @GetMapping
     public String getAdminReservations() {
         return "/admin/reservation";
-    }
-
-    @GetMapping("/reservations")
-    public ResponseEntity<List<ReservationResponse>> getReservations() {
-        List<ReservationResponse> responses = reservationService.findAllReservations();
-
-        return ResponseEntity.ok(responses);
-    }
-
-    @PostMapping("/reservations")
-    public ResponseEntity<ReservationResponse> createReservation(@RequestBody ReservationSaveRequest request) {
-        ReservationResponse response = reservationService.saveReservation(request);
-
-        URI location = URI.create("/reservations/" + response.id());
-        return ResponseEntity.created(location).body(response);
-    }
-
-    @DeleteMapping("/reservations/{id}")
-    public ResponseEntity<Void> deleteReservation(@PathVariable Long id) {
-        reservationService.deleteReservationById(id);
-
-        return ResponseEntity.noContent().build();
     }
 }
