+++ src/main/java/roomescape/controller/ReservationRestController.java
@@ -1,0 +1,43 @@
+package roomescape.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import roomescape.dto.ReservationResponse;
+import roomescape.dto.ReservationSaveRequest;
+import roomescape.service.ReservationService;
+
+import java.net.URI;
+import java.util.List;
+
+@RestController
+@RequestMapping(path = "/reservations")
+public class ReservationRestController {
+
+    private final ReservationService reservationService;
+
+    public ReservationRestController(ReservationService reservationService) {
+        this.reservationService = reservationService;
+    }
+
+    @GetMapping
+    public ResponseEntity<List<ReservationResponse>> getReservations() {
+        List<ReservationResponse> responses = reservationService.findAllReservations();
+
+        return ResponseEntity.ok(responses);
+    }
+
+    @PostMapping
+    public ResponseEntity<ReservationResponse> createReservation(@RequestBody ReservationSaveRequest request) {
+        ReservationResponse response = reservationService.saveReservation(request);
+
+        URI location = URI.create("/reservations/" + response.id());
+        return ResponseEntity.created(location).body(response);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteReservation(@PathVariable Long id) {
+        reservationService.deleteReservationById(id);
+
+        return ResponseEntity.noContent().build();
+    }
+}
